class-theme-my-login-ms-signup.php HACK
By commenting out these files, the redirection is avoided and the
registration is successfully done in the sub-site.  If WP MS is being
used as a site with different languages, this would mean the user gets
attached to the correct site for his language and all welcome emails
and further actions are taken in that language.